requirements.txt
google-cloud-vision>=3.4.0
sympy>=1.12
pillow>=10.0.0